+++ CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.5.0](https://github.com/lerna/lerna/compare/v5.4.3...v5.5.0) (2022-08-31)
+
+
+### Bug Fixes
+
+* **version:** only update existing lockfile deps ([#3308](https://github.com/lerna/lerna/issues/3308)) ([f5c8480](https://github.com/lerna/lerna/commit/f5c84807b957079b6a6d771bb9ca327f88c15a8a))
+
+
+### Features
+
+* pnpm workspaces support ([#3284](https://github.com/lerna/lerna/issues/3284)) ([1b18dbe](https://github.com/lerna/lerna/commit/1b18dbeb03e98c70b5428a9fe457781a59f8d65d))
+
+
+
 ## [5.4.3](https://github.com/lerna/lerna/compare/v5.4.2...v5.4.3) (2022-08-16)